Fix compile warning in Inja

+++ external/html/inja/inja.hpp
@@ -2159,7 +2159,7 @@ class Renderer : public NodeVisitor {
 
   static std::string htmlescape(const std::string& data) {
     std::string buffer;
-    buffer.reserve(1.1 * data.size());
+    buffer.reserve(static_cast<std::size_t>(1.1 * data.size()));
     for (size_t pos = 0; pos != data.size(); ++pos) {
       switch (data[pos]) {
         case '&':  buffer.append("&amp;");       break;
